+++ tgui/packages/tgui/interfaces/LibraryAdmin/BookListing.tsx
@@ -2,7 +2,8 @@ import { Box, NoticeBox, Stack } from 'tgui-core/components';
 
 import { useBackend } from '../../backend';
 import { Window } from '../../layouts';
-import { PageSelect, SearchAndDisplay } from '../LibraryConsole/components';
+import { PageSelect } from '../LibraryConsole/components/PageSelect';
+import { SearchAndDisplay } from '../LibraryConsole/components/Search';
 import { LibraryAdminData } from './types';
 
 export function BookListing(props) {
