+++ tgui/packages/tgui/interfaces/LibraryConsole/components/PageSelect.tsx
@@ -1,19 +1,4 @@
-import { Box, Button, Input, Section, Stack } from 'tgui-core/components';
-
-export function ScrollableSection(props) {
-  const { header, contents } = props;
-
-  return (
-    <Section fill scrollable>
-      <Box fontSize="20px" textAlign="center">
-        {header}
-      </Box>
-      <Box position="relative" top="10px">
-        {contents}
-      </Box>
-    </Section>
-  );
-}
+import { Button, Input, Stack } from 'tgui-core/components';
 
 export function PageSelect(props) {
   const {
