+++ tgui/packages/tgui/interfaces/LibraryConsole/components/Search.tsx
@@ -7,8 +7,8 @@ import {
   Table,
 } from 'tgui-core/components';
 
-import { useBackend } from '../../backend';
-import { LibraryConsoleData } from './types';
+import { useBackend } from '../../../backend';
+import { LibraryConsoleData } from '../types';
 
 export function SearchAndDisplay(props) {
   return (
