+++ tgui/packages/tgui/interfaces/LibraryVisitor.jsx
@@ -12,7 +12,7 @@ import {
 
 import { useBackend } from '../backend';
 import { Window } from '../layouts';
-import { PageSelect } from './LibraryConsole/components';
+import { PageSelect } from './LibraryConsole/components/PageSelect';
 
 export const LibraryVisitor = (props) => {
   return (
